refactor(rmem): promote inspector:rmem:* to top-level rmem:*
The .rbt-format tools (rbt:analyze / rbt:explore / rbt:recover)
live at the top level while the .rmem-format tools were sitting a
namespace deeper under inspector:rmem:*. Promote the rmem family
up so the two format-oriented command groups sit side-by-side:

  inspector:rmem:explore → rmem:explore
  inspector:rmem:live    → rmem:live
  inspector:rmem:mcp     → rmem:mcp
  inspector:rmem:query   → rmem:query
  inspector:rmem:serve   → rmem:serve
  inspector:rmem:viz     → rmem:viz

The command classes move from Reli\Command\Inspector\ to
Reli\Command\Rmem\ to match the directory-based CommandEnumerator.
README, docs and cross-command references (TextReportFormatter's
"Explore: …" hints) are rewritten to the new names; the psalm
baseline file paths follow the move; the MCP command's integration
test moves to tests/Command/Rmem/ and picks up the new namespace.

This feature set has not been in a tagged release yet so no backwards
aliases are kept — callers updating from the branch tip need to
replace inspector:rmem: with rmem: everywhere.
