+++ src/Command/Rmem/RmemQueryCommand.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Reli\Command\Inspector;
+namespace Reli\Command\Rmem;
 
 use Reli\Inspector\Output\MemoryOutput\BinaryFormat\Format;
 use Reli\Inspector\Output\MemoryOutput\BinaryFormat\Reader as BinaryReader;
@@ -30,7 +30,7 @@ final class RmemQueryCommand extends Command
     #[\Override]
     public function configure(): void
     {
-        $this->setName('inspector:rmem:query')
+        $this->setName('rmem:query')
             ->setDescription('Query a .rmem file for node details and tree paths')
             ->addArgument('file', InputArgument::OPTIONAL, '.rmem file path (not needed with --server)')
             ->addOption(
